Add tag type
This is just a nice-to-have feature, colors tags in the tag viewer
by type.

I decided to handle the character tags by pulling a list of the 250
most popular character tags from Derpibooru, and checking against
those. This covers every character with over 600 images. This makes
the client subtly incompatible with any other site that might use
the Derpibooru API, but I don't think that's a significant concern
at this time.